requirements.txt
flask
flask-sqlalchemy
geopandas
pyogrio
shapely
fiona
python-dotenv
sqlalchemy-utils
mapbox-vector-tile
gunicorn
